+++ src/Resources/DbSchemaResource.php
@@ -169,7 +169,7 @@ class DbSchemaResource extends BaseDbResource
     {
         if ($refresh || (is_null($tables = $this->parent->getFromCache('tables')))) {
             $tables = [];
-            foreach ($this->parent->getSchemas() as $schemaName) {
+            foreach ($this->parent->getSchemas($refresh) as $schemaName) {
                 $result = $this->parent->getSchema()->getResourceNames(DbResourceTypes::TYPE_TABLE, $schemaName);
                 $tables = array_merge($tables, $result);
                 // Until views are separated as separate resource